Fix ComponentButtons
Enable the ButtonGroup in ComponentButtons.
Fix opening and closing of FullDetailPage for component inspection
in PageDefinition (workspace).

ComponentButtons is a common component used in PageBrowse,
PageDefinition, and PageCuration.
Some of the buttons do not make sense in a certain view:
-PageBrowse does not render FullDetailPage, so onInspect, which make
the contained FullDetailPage visible does not make sense there.
-onAddComponent is implemented in UserManagedList.  PageBrowse is
not a subclass of UserManagedList and therefore will not have
onAddComponent available.  Adding (source) component should then be
disabled in PageBrowse

Task: https://github.com/clearlydefined/website/issues/946