Fix scrolling in ComponentList
1. The height on the clearly-table-body crops the scroll bar, so
scrolling cannot reach the item at the end of the list.  Moved the
height styling to address this.

2. There is a blank space after components in the table even when
there is not enough space to display all the components.  This is
because renderFilterBar, passed as props from the container components
(e.g. PageDefinition), is not called in ComponentList.  Padding to
accommodate the FilterBar is added by the parent of the ComponentList,
thus leaving a blank space after the ComponentList when FilterBar is
not rendered inside ComponentList.

Change the spacing from padding in the section (parent) to margin of
the ComponentList (child), where showing filter bar is determined.
Also add the conditional formatting to control the spacing based on
whether filter bar is displayed.